Replace AltC with BTreeC.
