+++ src/Control/Effect.hs
@@ -9,7 +9,7 @@ import Control.Effect.Error     as X (Error, ErrorC)
 import Control.Effect.Fail      as X (Fail, FailC)
 import Control.Effect.Fresh     as X (Fresh, FreshC)
 import Control.Effect.Lift      as X (Lift, LiftC, runM)
-import Control.Effect.NonDet    as X (NonDet, AltC)
+import Control.Effect.NonDet    as X (NonDet, BTreeC)
 import Control.Effect.Random    as X (Random, RandomC)
 import Control.Effect.Reader    as X (Reader, ReaderC)
 import Control.Effect.Resource  as X (Resource, ResourceC)
